remove setup_env call from import
this mechanism was designed to make configuring source code and data directories easier, but in practice it makes things more difficult everywhere except for a couple of UC GPU clusters. the user is still free to set up and use the nugraph environment variables, but they are no longer a condition for importing the module.